be/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
asyncpg==0.29.0
